[#72-search-categoryId] 조회에 카테고리ID 통일
