+++ build.gradle
@@ -80,7 +80,7 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-//	testImplementation 'org.springframework.security:spring-security-test'
+	testImplementation 'org.springframework.security:spring-security-test'
 
 	implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
