--- src/test/resources/test.sql
@@ -1,1 +1,0 @@
-select 1;
